modify: move SL egendaries to talents
