+++ src/analysis/retail/paladin/holy/modules/talents/MaraadsCastRatio.tsx
@@ -22,7 +22,7 @@ class MaraadsCastRatio extends Analyzer {
 
   constructor(options: Options) {
     super(options);
-    this.active = this.selectedCombatant.hasTalent(TALENTS.MARAADS_DYING_BREATH_HOLY_TALENT);
+    this.active = this.selectedCombatant.hasTalent(TALENTS.MARAADS_DYING_BREATH_TALENT);
     if (!this.active) {
       return;
     }
